# Use expm1 and general optimizations