Use expm1 instead to improve accuracy. Also, optimized several other places.

```julia
# before
julia> @benchmark fluence_DA_Nlay_cylinder_CW(1.0, (0.2, 0.1, 0.15, 0.14, 0.3), (10.0, 12.0, 11.0, 11.5, 14.2), n_med=(1.0, 1.0, 1.0, 1.0, 1.0), l=(0.2, 0.2, 0.2, 0.2, 2.0), MaxIter=5000, atol=1e-300)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  360.812 μs … 524.975 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     365.948 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   366.296 μs ±   3.620 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▃▄▁       ▆█▇▄  ▃▄▂▂▁                                      ▂
  ▃▅▆███▄▄▆▇▇▅▇█████▆█████▇▆▄▅▆▅▄▇▇▄▅▅▃▅▅▅▄▁▃▆▆▆▆▄▄▁▁▄▄▄▅▄▅▅▅▆▇ █
  361 μs        Histogram: log(frequency) by time        382 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

# after
julia> @benchmark fluence_DA_Nlay_cylinder_CW(1.0, (0.2, 0.1, 0.15, 0.14, 0.3), (10.0, 12.0, 11.0, 11.5, 14.2), n_med=(1.0, 1.0, 1.0, 1.0, 1.0), l=(0.2, 0.2, 0.2, 0.2, 2.0), MaxIter=5000, atol=1e-300)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  296.768 μs … 673.499 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     298.782 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   300.220 μs ±  15.265 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

        ▅█  ▇▄                                                   
  ▂▂▁▁▂▃██▆▆██▅▃▂▃▅▅▄▃▃▄▄▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▂▂▂▂▂▂▂▁▂▂▂▂▂▂ ▃
  297 μs           Histogram: frequency by time          308 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```